feat: improved css for sign-in | removed register code from sign-in
CSS was improved via a changing the structure to be easier to understand and also via adding better commnets/more comments in the css file.

Removed register code from the html file of the sign-in view.

Register code/styling was removed from both the css file due to it no longer having a use.